[#116752703] Done: As an user, I'd like to dismiss the badge on Chat tab and UNREAD MESSAGES once I launch the keyboard

+++ meWrap/Constants.swift
@@ -22,7 +22,7 @@ class Constants: NSObject {
     static let wrapNameLimit = 190
     static let recentCandiesLimit = 6
     static let recentCandiesLimit_2 = 3
-    static let composeBarDefaultCharactersLimit: CGFloat = 21000
+    static let composeBarDefaultCharactersLimit: Int = 21000
     static let encryptedAuthorization = "encrypted_authorization"
     static let groupIdentifier = "group.com.ravenpod.wraplive"
 }
